refactor: 에러메시지 객체를 도메인별로 분리
