+++ src/main/java/racingcar/domain/error/car/RacingCarErrorMessage.java
@@ -1,0 +1,6 @@
+package racingcar.domain.error.car;
+
+public class RacingCarErrorMessage {
+    public final static String DUPLICATE_NAME = "차 이름은 중복될 수 없습니다.";
+    public final static String OVER_NAME_LENGTH = "차 이름 길이의 제한을 넘었습니다.";
+}
