+++ src/main/java/racingcar/domain/error/racing/RacingErrorMessage.java
@@ -1,6 +1,6 @@
-package racingcar.domain.error;
+package racingcar.domain.error.racing;
 
-public class ErrorMessage {
+public class RacingErrorMessage {
     public final static String INVALID_INPUT_TEXT = "잘못된 입력입니다.";
     public final static String INVALID_TRY_COUNT = "잘못된 시도횟수 입력입니다.";
 }
